+++ src/index.ts
@@ -34,13 +34,24 @@ export {
   setExtras,
   setTag,
   setTags,
+  flush,
   setUser,
   // eslint-disable-next-line deprecation/deprecation
   startTransaction,
+  withScope,
+
+  // V8
+  startInactiveSpan,
   startSpan,
   startSpanManual,
-  startInactiveSpan,
-  withScope,
+  getActiveSpan,
+  spanToJSON,
+  spanIsSampled,
+  setMeasurement,
+  getGlobalScope,
+  // getIsolationScope, TODO: Verify why it's not working
+  setCurrentClient,
+
 } from '@sentry/core';
 export { Replay, BrowserTracing } from '@sentry/browser'